Add files via upload
Data strucure:- A data structure is a way to store data. We structure data in different ways depending on what data we have, and what we want to do with it.Data structures give us the possibility to manage large amounts of data efficiently for uses such as large databases and internet indexing services.Data structures are essential ingredients in creating fast and powerful algorithms. They help in managing and organizing data, reduce complexity, and increase efficiency.

there are two different kinds of data structures.

Primitive Data Structures are basic data structures provided by programming languages to represent single values, such as integers, floating-point numbers, characters, and booleans.

Abstract Data Structures are higher-level data structures that are built using primitive data types and provide more complex and specialized operations. Some common examples of abstract data structures include arrays, linked lists, stacks, queues, trees, and graphs.

Stacks:-A stack is a data structure that can hold many elements.

In a pile of pancakes, the pancakes are both added and removed from the top. So when removing a pancake, it will always be the last pancake you added. This way of organizing elements is called LIFO: Last In First Out.

Basic operations we can do on a stack are:

Push: Adds a new element on the stack.

Pop: Removes and returns the top element from the stack.

Peek: Returns the top element on the stack.

isEmpty: Checks if the stack is empty.

Size: Finds the number of elements in the stack.

Queues:- A queue is a data structure that can hold many elements.

Think of a queue as people standing in line in a supermarket.

The first person to stand in line is also the first who can pay and leave the supermarket. This way of organizing elements is called FIFO: First In First Out.

Basic operations we can do on a queue are:

Enqueue: Adds a new element to the queue.

Dequeue: Removes and returns the first (front) element from the queue.

Peek: Returns the first element in the queue.

isEmpty: Checks if the queue is empty.

Size: Finds the number of elements in the queue.

Linked Lists:- A linked list consists of nodes with some sort of data, and a pointer, or link, to the next node.

There are three basic forms of linked lists:

Singly linked lists:-A singly linked list is the simplest kind of linked lists. It takes up less space in memory because each node has only one address to the next node

Doubly linked lists:-A doubly linked list has nodes with addresses to both the previous and the next node, like in the image below, and therefore takes up more memory. But doubly linked lists are good if you want to be able to move both up and down in the list.

Circular linked lists:-A circular linked list is like a singly or doubly linked list with the first node, the "head", and the last node, the "tail", connected.

Trees:- The Tree data structure is similar to Linked Lists in that each node contains data and can be linked to other nodes. We have previously covered data structures like Arrays, Linked Lists, Stacks, and Queues. These are all linear structures, which means that each element follows directly after another in a sequence. Trees however, are different. In a Tree, a single element can have multiple 'next' elements, allowing the data structure to branch out in various directions.

Types of Trees Trees are a fundamental data structure in computer science, used to represent hierarchical relationships. This tutorial covers several key types of trees.

Binary Trees: Each node has up to two children, the left child node and the right child node. This structure is the foundation for more complex tree types like Binay Search Trees and AVL Trees.

Binary Search Trees (BSTs): A type of Binary Tree where for each node, the left child node has a lower value, and the right child node has a higher value.

AVL Trees: A type of Binary Search Tree that self-balances so that for every node, the difference in height between the left and right subtrees is at most one. This balance is maintained through rotations when nodes are inserted or deleted.

Hash Table:-A Hash Table is a data structure designed to be fast to work with. The reason Hash Tables are sometimes preferred instead of arrays or linked lists is because searching for, adding, and deleting data can be done really quickly, even for large amounts of data.

Linear search:- In Linear Search, we iterate over all the elements of the array and check if it the current element is equal to the target element. If we find any element to be equal to the target element, then return the index of the current element. Otherwise, if no element is equal to the target element, then return -1 as the element is not found. Linear search is also known as sequential search.

Binary search:- Binary search is a search algorithm used to find the position of a target value within a sorted array. It works by repeatedly dividing the search interval in half until the target value is found or the interval is empty. The search interval is halved by comparing the target element with the middle value of the search space.

Selection Sort:- The Selection Sort algorithm finds the lowest value in an array and moves it to the front of the array.

Bubble Sort:- Bubble Sort is an algorithm that sorts an array from the lowest value to the highest value.

Insertion Sort:- The Insertion Sort algorithm uses one part of the array to hold the sorted values, and the other part of the array to hold values that are not sorted yet.

Merge Sort:- The Merge Sort algorithm is a divide-and-conquer algorithm that sorts an array by first breaking it down into smaller arrays, and then building the array back together the correct way so that it is sorted.

Quicksort:- The Quicksort algorithm takes an array of values, chooses one of the values as the 'pivot' element, and moves the other values so that lower values are on the left of the pivot element, and higher values are on the right of it